+++ app/admin-new/(style)/themeInfo.modules.sass
@@ -177,7 +177,7 @@
         .hint-content
           color: white
           @include body14R
-          padding: 0 0 6px 4px
+          padding: 0 0 6px 0
         .hint-images
           display: flex
         .hint-image-box
